CRUD semi completo, falta findbyid.

+++ API/Data/VO/ItemVO.cs
@@ -10,8 +10,8 @@ namespace API.Data.VO
     {
         public long Codigo { get; set; }
         public string Descricao { get; set; }
-        public double Estoque { get; set; }
-        public double Preco { get; set; }
+        public decimal Estoque { get; set; }
+        public decimal Preco { get; set; }
         public DateTime Cadastro { get; set; }
         public bool Ativo { get; set; }
     }
